Indicators of Electoral Systems and Party Systems

It allows structuring electoral data of different size and structure to calculate various indicators frequently used in the studies of electoral systems and party systems:

Electoral Volatility:

Pedersen (1979) <doi:10.1111/j.1475-6765.1979.tb01267.x>

Powell and Tucker (2014) <doi:10.1017/S0007123412000531>

Torcal and Lago (2015, ISBN:9788415260356)

Effective Number of Parties:

Laakso and Taagepera (1979) <doi:10.1177/001041407901200101>

Electoral disproportionality:

Rae (1971, ISBN:9780300015171)

Loosemore and Hanby (1971) <doi:10.1017/S000712340000925X>

Lijphart (1986, ISBN:0875860648, 0875860745)

Lijphart (1994) <doi:10.1093/acprof:oso/9780198273479.001.0001>

Cox and Shugart (1991) <doi:10.1016/0261-3794(91)90025-N>

Gallagher (1991) <doi:10.1016/0261-3794(91)90004-C>

Party System Nationalization:

Mainwaring and Jones (2003) <doi:10.1177/13540688030092002>

Chhibber and Kollman (2004, ISBN:9781400826377)

See also