Party System Nationalization Score

Party System Nationalization Score Mainwaring and Jones (2003) and Chhibber and Kollman (2004)

psns(tidy_data, method, pns = FALSE, scale = 100)

Arguments

tidy_data
data.frame that contains the following variables with these names:

election: year of election.

unit: the unit of analysis (province, department ...)

party: name of the political parties that obtain votes.

votes: votes obtained by each party.

votes_nac: votes at national level for each party.

If the data is not structured in this way you can order it with: convert_esaps.

method
Method to calculate Party System Nationalization Score:

method = "Mainwaring and Jones" or method = 1.

method = "Chhibber and Kollman" or method = 2.

pns
by default it is FALSE. If TRUE, the Party Nationalization Score is calculated. In method, you must indicate: method = 1.

scale
By default it is 100, the indices will vary between 0 and 100. If scale = 1 the variation will be between 0 and 1.

Value

if pns = FALSE, return data.frame.

if pns = TRUE, return a list with two data.frame.

list[[1]] PSNS: Party System Nationalization Score

list[[2]] PNS: Party Nationalization Score

Author

Nicolas Schmidt nschmidt@cienciassociales.edu.uy

Examples

votes <- data.frame(election = rep(2000,4),
                    unit  = rep(c("District_1", "District_2"), each = 2),
                    party = rep(c("party_A", "party_B"), 2),
                    votes = c(0.60,0.40, 0.30, 0.70),
                    votes_nac = rep(c(0.55,0.45),2)
                   )
psns(tidy_data = votes, method = 1)

#>   election   psns
#> 1     2000 84.695

psns(tidy_data = votes, method = 1, pns = TRUE)

#> $PSNS
#>   election   psns
#> 1     2000 84.695
#> 
#> $PNS
#>     party election   pns
#> 1 party_A     2000 0.833
#> 2 party_B     2000 0.864
#>